requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
colorama==0.4.5
crashtest==0.3.1
cryptography==37.0.4
distlib==0.3.6
dnspython==2.2.1
dulwich==0.20.46
email-validator==1.2.1
fastapi==0.82.0
filelock==3.8.0
h11==0.13.0
html5lib==1.1
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
jaraco.classes==3.2.2
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.1
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
orjson==3.8.0
packaging==21.3
pexpect==4.8.0
pipenv==2022.9.8
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==1.10.2
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
